Crate stm_util

A hardware abstraction layer for the STM32F070CBT6 MCU for use in the calc-rs project

Modules

address

gpio

i2c

prelude

timer

util

Structs

CriticalSection

Functions

clear_pending_interrupt⚠

Clears the pending status of the interupt with an IRQ number interrupt. This function does nothing if the interrupt is not pending Safety: interrupt must be in the range 0..32, and there must be a valid interrupt in the interrupt vector table

data_memory_barrier

Waits until previous memory accesses are done before doing any more memory accesses

disable_interrupts

enable_interrupt⚠

Enabled the interrupt with the number interrupt Safety: interrupt must be in the range 0..32, and there must be a valid interrupt in the interrupt vector table

enable_interrupts

get_primask

wait_for_interrupt